--- custom_components/hacs/models/lovelace.py
@@ -1,10 +1,0 @@
-"""Lovelace Modes."""
-from enum import Enum
-
-
-class LovelaceMode(str, Enum):
-    """Lovelace Modes."""
-
-    STORAGE = "storage"
-    AUTO = "auto"
-    YAML = "yaml"
