--- custom_components/hacs/models/stage.py
@@ -1,10 +1,0 @@
-"""Helper constants."""
-# pylint: disable=missing-class-docstring
-from enum import Enum
-
-
-class HacsStage(str, Enum):
-    SETUP = "setup"
-    STARTUP = "startup"
-    RUNNING = "running"
-    BACKGROUND = "background"
