+++ custom_components/hacs/repositories/appdaemon.py
@@ -3,7 +3,7 @@ from aiogithubapi import AIOGitHubAPIException
 
 from custom_components.hacs.helpers.classes.exceptions import HacsException
 from custom_components.hacs.helpers.classes.repository import HacsRepository
-from custom_components.hacs.models.category import HacsCategory
+from custom_components.hacs.enums import HacsCategory
 from custom_components.hacs.helpers.functions.logger import getLogger
