+++ custom_components/hacs/repositories/theme.py
@@ -1,7 +1,7 @@
 """Class for themes in HACS."""
 from custom_components.hacs.helpers.classes.exceptions import HacsException
 from custom_components.hacs.helpers.classes.repository import HacsRepository
-from custom_components.hacs.models.category import HacsCategory
+from custom_components.hacs.enums import HacsCategory
 from custom_components.hacs.helpers.functions.information import find_file_name
 from custom_components.hacs.helpers.functions.logger import getLogger
